Count ballots not sample draws for opportunistic contests
Previously, we counted votes for targeted and opportunistic contests the
same way - using the sample draws. This was incorrect. For targeted
contests, we should count the sample draws. For opportunistic, we should
count distinct ballots.

This will make more a difference when we have multiple contests with
distinct sample draws that may refer to the same ballots.